Lint

+++ programs/helium-sub-daos/src/instructions/delegation/claim_rewards_v0.rs
@@ -138,7 +138,7 @@ pub fn handler(ctx: Context<ClaimRewardsV0>, args: ClaimRewardsArgsV0) -> Result
   // calculate the position's share of that epoch's rewards
   // rewards = staking_rewards_issued * staked_vehnt_at_epoch / total_vehnt
   let rewards = u64::try_from(
-    (delegated_vehnt_at_epoch as u128)
+    delegated_vehnt_at_epoch
       .checked_mul(ctx.accounts.sub_dao_epoch_info.delegation_rewards_issued as u128)
       .unwrap()
       .checked_div(ctx.accounts.sub_dao_epoch_info.vehnt_at_epoch_start as u128)
